Rename Bo pages
