+++ tests/puppeteer/campaigns/functional/BO/catalog/brandsAndSuppliers/suppliers/01_CRUDSupplier.js
@@ -8,10 +8,10 @@ const SupplierFaker = require('@data/faker/supplier');
 const BOBasePage = require('@pages/BO/BObasePage');
 const LoginPage = require('@pages/BO/login');
 const DashboardPage = require('@pages/BO/dashboard');
-const BrandsPage = require('@pages/BO/brands');
-const SuppliersPage = require('@pages/BO/suppliers');
-const AddSupplierPage = require('@pages/BO/addSupplier');
-const ViewSupplierPage = require('@pages/BO/viewSupplier');
+const BrandsPage = require('@pages/BO/catalog/brands');
+const SuppliersPage = require('@pages/BO/catalog/suppliers');
+const AddSupplierPage = require('@pages/BO/catalog/suppliers/add');
+const ViewSupplierPage = require('@pages/BO/catalog/suppliers/view');
 
 let browser;
 let page;
