+++ tests/puppeteer/pages/BO/catalog/suppliers/add.js
