+++ tests/puppeteer/pages/BO/catalog/suppliers/index.js
