+++ tests/puppeteer/pages/BO/catalog/suppliers/view.js
